+++ src/components/Table/index.ts
@@ -9,4 +9,4 @@ export { ThCheckbox } from './ThCheckbox'
 export { TdCheckbox } from './TdCheckbox'
 export { BulkActionRow } from './BulkActionRow'
 export { EmptyTableBody } from './EmptyTableBody'
-export { FixColumnTable } from './FixColumnTable'
+export { TableReel } from './TableReel'
